+++ src/main/java/com/valetparkingtracker/enterprise/controllers/TicketController.java
@@ -1,0 +1,4 @@
+package com.valetparkingtracker.enterprise.controllers;
+
+public class TicketController {
+}
