+++ src/main/java/com/valetparkingtracker/enterprise/controllers/ValetParkingTrackerController.java
@@ -1,0 +1,71 @@
+package com.valetparkingtracker.enterprise.controllers;
+
+import com.valetparkingtracker.enterprise.dto.Vehicle;
+import com.valetparkingtracker.enterprise.service.IVehicleService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.*;
+
+/**
+ * The controller for Valet Parking Tracker REST endpoints and web UI
+ * <p>
+ *     This class handles the CRUD operations for Valet Parking Tracker vehicles through HTTP actions
+ * </p>
+ * <p>
+ *     This class also serves HTML pages for UI interactions with vehicles
+ * </p>
+ */
+@Controller
+public class ValetParkingTrackerController {
+
+    @Autowired
+    IVehicleService vehicleService;
+
+    /**
+     * Handle the root (/) endpoint and return a start page.
+     * @return
+     */
+    @RequestMapping("/")
+    public String index(Model model) {
+        Vehicle vehicle = new Vehicle();
+        vehicle.setVehicleId("25");
+        vehicle.setMake("VW");
+        vehicle.setModel("Golf");
+        vehicle.setColor("Silver");
+        vehicle.setNotes("damaged");
+        vehicle.setTicketId("27");
+        model.addAttribute(vehicle);
+        return "start";
+    }
+
+    @RequestMapping("/saveVehicle")
+    public String saveVehicle(Vehicle vehicle) {
+        try {
+            vehicleService.save(vehicle);
+        } catch (Exception e) {
+            e.printStackTrace();
+            return "start";
+        }
+        return "start";
+    }
+
+    /**
+     * <p>UI Mapping</p>
+     *
+     * Displays the form to check in a guest
+     * Should create a new vehicle, customer, and ticket using provided info
+     *
+     * @return the newVehicle html page
+     */
+    @RequestMapping("/checkIn")
+    public String checkIn(Model model) {
+        Vehicle vehicle = new Vehicle();
+        vehicle.setMake("VW");
+        vehicle.setModel("Golf");
+        vehicle.setColor("Silver");
+        vehicle.setNotes("damaged");
+        model.addAttribute(vehicle);
+        return "checkIn";
+    }
+}
